Add index package
This commit adds a new package that wraps up the existing indexing code
in a nicer API. I also separated out indexing the pages and posts for
this package without the addition of the configuration.

There are some notable changes to how pages and posts have been loaded,
most significantly in that the contents are no longer stored in memory.
With the anticipation that this will quickly cause memory issues as a
blog grows, I now only store the paths to the files which can be used
during static generation or in the live server to load the file and
render it lazily.