Refactor post indexing
This commit rethinks how post indexing works in concert with the other
components of the project, settling on an approach that relies on strict
invariants rather than hidden behavior.

Now, rather than attempting to sort a list of posts itself when
generating a feed, the rss package will rely on the posts being
pre-sorted. In order to support this invariant, the index.New function
has been updated to guarantee an ordered set of posts.

The sort order of posts is not configurable, but this still leaves the
ability for the package user to sort them in whatever way makes sense.
In addition, if a user does not want to use the provided indexing code
they can still benefit from the rss package, they will simply need to
ensure that invariants are met themselves.

This feels like the best trade off between functionality and simplicity
considering the entire goal of the project is to minimize configuration.