fix: Correct frontmatter parsing
This commit modifies the markdown/ParseFrontmatter function to fix an
issue with frontmatter not being populated correctly. The underlying
issue was that the function takes an any type which we were
dereferencing when passing it to the Decode method. Since we're already
passing a reference to the data struct, we can omit the extra reference
which fixes the bug.