feat: Add support for draft posts
This commit adds a new `draft` field to the post frontmatter to allow
users to author posts that show in the local development server, but are
not included in a production build. This new field is accompanied by a
new `date` field that replaces the previous filename-based timestamp
parsing.

This latter change was made to give us the ability to easily add a
business rule that ensures that draft posts don't have an attached
timestamp. This was done to avoid the situation where a draft was given a
timestamp at the start of authoring, but was not updated when the draft
flag was removed resulting in a post that renders way in the past.